+++ frontend/packages/react-app/src/components/molecules/DistributionTargetList/index.stories.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import { Story, Meta } from "@storybook/react/types-6-0";
 import { BrowserRouter } from "react-router-dom";
 import DistributionTargetList, { TargetsProps } from "./index";
-import { Target } from "../../../interfaces";
+import { distributionTargets } from "../../../utils/mockData";
 
 export default {
   title: "Molecules/DistributionTargetList",
@@ -15,35 +15,12 @@ const Template: Story<TargetsProps> = (args) => (
   </BrowserRouter>
 );
 
-const targets: Target[] = [
-  {
-    handle: "sample user",
-    wallet: "0x4B8619890fa9C3cF11C497961eB4b970D440127F",
-  },
-  {
-    handle: "username sample user",
-    wallet: "0x4B8619890fa9C3cF11C497961eB4b970D440127F",
-  },
-  {
-    handle: "sample",
-    wallet: "0x4B8619890fa9C3cF11C497961eB4b970D440127F",
-  },
-  {
-    handle: "username sample user",
-    wallet: "0x4B8619890fa9C3cF11C497961eB4b970D440127F",
-  },
-  {
-    handle: "sample user",
-    wallet: "0x4B8619890fa9C3cF11C497961eB4b970D440127F",
-  },
-];
-
 export const Default = Template.bind({});
 Default.args = {
-  targets,
+  distributionTargets,
 };
 
 export const NoTarget = Template.bind({});
 NoTarget.args = {
-  targets: [],
+  distributionTargets: [],
 };
