+++ frontend/packages/react-app/src/components/organisms/CampaignDetail/index.stories.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import { Story, Meta } from "@storybook/react/types-6-0";
 import { BrowserRouter } from "react-router-dom";
 import CampaignDetail, { CampaignDetailProps } from "./index";
-import { CampaignInfo } from "../../../interfaces";
+import { campaign } from "../../../utils/mockData";
 
 export default {
   title: "Organisms/CampaignDetail",
@@ -15,26 +15,7 @@ const Template: Story<CampaignDetailProps> = (args) => (
   </BrowserRouter>
 );
 
-const campaignInfo: CampaignInfo = {
-  campaignInfoCid: "QmYdUkdEgQtn4viySEQie51JAwrC3xTyDqDbdWXWEwAz5J",
-  campaignMetadata: {
-    name: "Audiusフォロワーさん向けトークン配布キャンペーン！",
-    description:
-      "Audiusアカウントのフォロワーの方へトークンを配布します。手順はXXX...",
-    image: "https://example.com/campaignimage.jpg",
-  },
-  claimAmount: "0",
-  distributor: {
-    id: "0x48889feca4574810e5a5b30b6b93146a837500fb",
-  },
-  id: "0xcc22b3199013627708d05fedf9b854bd7706a156",
-  startDate: "20201212",
-  status: 0,
-  claimed: 10,
-  depositAmount: "100000",
-};
-
 export const Default = Template.bind({});
 Default.args = {
-  campaignInfo,
+  campaignInfo: campaign,
 };
