+++ frontend/packages/react-app/src/components/organisms/TokenList/index.stories.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import { Story, Meta } from "@storybook/react/types-6-0";
 import { BrowserRouter } from "react-router-dom";
 import TokenList, { TokenListProps } from "./index";
-import { TokenListState } from "../../../reducers/tokens";
+import { TokenListState } from "../../../interfaces";
 
 export default {
   title: "Organisms/TokenList",
