+++ frontend/packages/react-app/src/components/templates/ExplorePageTemplate/index.stories.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import { Story, Meta } from "@storybook/react/types-6-0";
 import { BrowserRouter } from "react-router-dom";
 import ExplorePageTemplate, { ExplorePageTemplateProps } from "./index";
-import { TokenListState } from "../../../reducers/tokens";
+import { tokenListState } from "../../../utils/mockData";
 
 export default {
   title: "Templates/ExplorePageTemplate",
@@ -15,45 +15,15 @@ const Template: Story<ExplorePageTemplateProps> = (args) => (
   </BrowserRouter>
 );
 
-const state: TokenListState = {
-  tokens: [
-    {
-      tokenAddress: "0xabcd....1234",
-      name: "Iroiro Token",
-      symbol: "IRO",
-      decimals: 8,
-      totalSupply: 2000000000,
-    },
-    {
-      tokenAddress: "0xabcd....1234",
-      name: "Storybook Token",
-      symbol: "STR",
-      decimals: 18,
-      totalSupply: 1000000000,
-    },
-    {
-      tokenAddress: "0xabcd....1234",
-      name: "Test Token",
-      symbol: "TST",
-      decimals: 4,
-      totalSupply: 3000000000,
-    },
-  ],
-  isOpen: false,
-  inputTokenAddress: "",
-  tokenAddress: "",
-  type: "explore",
-  color: "itblue",
-};
-
 export const Default = Template.bind({});
 Default.args = {
-  state,
+  state: tokenListState,
 };
 
 export const NoTokens = Template.bind({});
 NoTokens.args = {
   state: {
+    ...tokenListState,
     tokens: [],
     isOpen: false,
     inputTokenAddress: "",
@@ -66,6 +36,7 @@ NoTokens.args = {
 export const Loading = Template.bind({});
 Loading.args = {
   state: {
+    ...tokenListState,
     tokens: [],
     isOpen: false,
     inputTokenAddress: "",
